Dedicated module for TokenRepository
Moved the TokenRepository class out of the main app, and import a
singleton to fetch/store access tokens.
